--- a/COPYRIGHT
+++ b/COPYRIGHT
@@ -8,7 +8,7 @@ and commercial purposes at absolutely no cost.
 
 ===============================================================================
 
-Copyright (C) 2010-2016 YaweiZhang <yawei.zhang@foxmail.com>.
+Copyright (C) 2010-2017 YaweiZhang <yawei.zhang@foxmail.com>.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal