--- a/HISTORY
+++ b/HISTORY
@@ -123,6 +123,9 @@ HISTORY for Log4z
  *  check configure's checksum when auto update it.
  *  some other optimize.
  *
+ * VERSION 3.5 <DATE: 2017.08.20>
+ *  optimization
+ *
  */
    
 (end of HISTORY)