--- a/log4z.cpp
+++ b/log4z.cpp
@@ -144,9 +144,22 @@ public:
         _file = fopen(path, mod);
         return _file != NULL;
     }
+    inline void clean(int index, int len)
+    {
+#if !defined(__APPLE__) && !defined(WIN32) 
+       if (_file != NULL)
+       {
+          int fd = fileno(_file);
+          posix_fadvise(fd, index, len, POSIX_FADV_DONTNEED);
+          fsync(fd);
+          posix_fadvise(fd, 0, 0, POSIX_FADV_DONTNEED);
+          fsync(fd);
+       }
+#endif
+    }
     inline void close()
     {
-        if (_file != NULL){fclose(_file);_file = NULL;}
+        if (_file != NULL){clean(0, 1000*1000*1000); fclose(_file);_file = NULL;}
     }
     inline void write(const char * data, size_t len)
     {