### 优化后
##### 优化内容  
* 整形/浮点数转换成字符串, 采用lut表方案, 替换原先的sprintf  * 去掉每次写日志的时间戳到tm检测, 以及每次写入文件时的sameday检测  
* 手写时间函数的序列化 
* 代码细节的优化调整  


单线程简单日志写入
```
Write Speed: 2240293 n/s, Speed: 224612 KB/s, Waiting: 18397
Write Speed: 2250669 n/s, Speed: 225653 KB/s, Waiting: 20894
Write Speed: 2216420 n/s, Speed: 222219 KB/s, Waiting: 21356
```

单线程 复杂类型写入  
```
Write Speed: 804541 n/s, Speed: 306940 KB/s, Waiting: 7276
Write Speed: 812761 n/s, Speed: 310076 KB/s, Waiting: 20609
Write Speed: 820467 n/s, Speed: 313016 KB/s, Waiting: 8216  
```

4线程复杂日志写入  
```
Write Speed: 124934 n/s, Speed: 47755 KB/s, Waiting: 20642
Write Speed: 129670 n/s, Speed: 49565 KB/s, Waiting: 25567
Write Speed: 129472 n/s, Speed: 49490 KB/s, Waiting: 28734
```


### 优化前  
单线程复杂日志写入  
```
Write Speed: 136955 n/s, Speed: 51357 KB/s, Waiting: 11424
Write Speed: 140889 n/s, Speed: 52833 KB/s, Waiting: 7108
Write Speed: 140706 n/s, Speed: 52764 KB/s, Waiting: 2607
 ```

单线程简单日志写入  
```
Write Speed: 288228 n/s, Speed: 26739 KB/s, Waiting: 12330
Write Speed: 285721 n/s, Speed: 26507 KB/s, Waiting: 19592
Write Speed: 293157 n/s, Speed: 27197 KB/s, Waiting: 801 
```
4线程复杂日志写入  
```
Write Speed: 833111 n/s, Speed: 77290 KB/s, Waiting: 19864
Write Speed: 850984 n/s, Speed: 78948 KB/s, Waiting: 18405
Write Speed: 853380 n/s, Speed: 79171 KB/s, Waiting: 14672
```